move LoginController to production
